docs: sidenav fixes
also ports uxdot-sidenav elements to typescript
also loads global css on each page
also reduces dom payload for uxdot-sidenav elements

chore: fix serve command

docs: fix tokens sidenav
