+++ docs/_includes/layouts/base.njk
@@ -17,6 +17,7 @@
   <head>
     {% include "../partials/meta.html" %}
 
+    <link rel="stylesheet" href="/assets/packages/@rhds/tokens/css/global.css">
     <link rel="stylesheet" href="/styles/styles.css">
 
    {# On browsers that don't yet support native declarative shadow DOM, a
