+++ docs/assets/javascript/elements/uxdot-sidenav-dropdown-menu-item.css
@@ -1,0 +1,4 @@
+a {
+  padding: var(--rh-space-md) 0 var(--rh-space-md) var(--rh-space-lg);
+  font-size: var(--rh-font-size-body-text-md);
+}
