+++ docs/assets/javascript/elements/uxdot-sidenav-dropdown-menu.css
@@ -1,0 +1,4 @@
+:host {
+  display: block;
+  padding-inline-start: var(--rh-space-2xl);
+}
