+++ docs/assets/javascript/elements/uxdot-sidenav-dropdown.css
@@ -1,0 +1,3 @@
+:host {
+  z-index: var(--uxdot-sidenav-z-index, 102);
+}
