+++ docs/assets/javascript/elements/uxdot-sidenav-item.css
@@ -1,0 +1,22 @@
+:host,
+a {
+  display: block;
+}
+
+a {
+  padding: var(--rh-space-lg) var(--rh-space-2xl);
+  font-size: var(--rh-font-size-body-text-lg);
+  text-decoration: none !important;
+  color: var(--rh-color-text-primary-on-light) !important;
+  border-inline-start: var(--rh-border-width-lg) solid transparent;
+}
+
+a:hover {
+  background: var(--rh-color-surface-lighter);
+  border-inline-start-color: var(--rh-color-border-subtle-on-light);
+}
+
+a.active {
+  background: var(--rh-color-surface-lighter);
+  border-inline-start-color: var(--rh-color-accent-brand-on-light);
+}
