+++ docs/assets/javascript/elements/uxdot-sidenav-lightdom.css
@@ -1,7 +1,3 @@
-uxdot-sidenav > ul > li {
-  margin-block: 0;
-}
-
 uxdot-sidenav-dropdown > details > summary {
   display: flex;
   flex-direction: row;
@@ -13,13 +9,9 @@ uxdot-sidenav-dropdown > details > summary {
     solid
     transparent;
   text-decoration: none;
-  color: var(--rh-color-text-primary-on-light, #151515);
-  font-size: var(--rh-font-size-body-text-lg, 1.125rem);
-  padding: var(--rh-space-lg, 16px) var(--rh-space-2xl, 32px);
-}
-
-uxdot-sidenav-dropdown-menu > ul > li {
-  margin-block: 0;
+  color: var(--rh-color-text-primary-on-light);
+  font-size: var(--rh-font-size-body-text-lg);
+  padding: var(--rh-space-lg) var(--rh-space-2xl);
 }
 
 uxdot-sidenav-dropdown > details > summary::-webkit-details-marker {
@@ -29,24 +21,18 @@ uxdot-sidenav-dropdown > details > summary::-webkit-details-marker {
 uxdot-sidenav-dropdown > details > summary:hover,
 uxdot-sidenav-dropdown > details[open] > summary:focus,
 uxdot-sidenav-dropdown > details[open] > summary:hover {
-  background: var(--rh-color-surface-lighter, #f2f2f2);
-  border-inline-start-color: var(--rh-color-border-subtle-on-light, #c7c7c7);
+  background: var(--rh-color-surface-lighter);
+  border-inline-start-color: var(--rh-color-border-subtle-on-light);
 }
 
 uxdot-sidenav-dropdown > details > summary:after {
   content: "";
-  color: var(--rh-color-text-primary-on-light, #151515);
+  color: var(--rh-color-text-primary-on-light);
   width: 9px;
   height: 9px;
   transform: rotate(45deg);
-  border-right:
-    var(--rh-border-width-lg, 3px)
-    solid
-    var(--rh-color-border-strong-on-light, #151515);
-  border-top:
-    var(--rh-border-width-lg, 3px)
-    solid
-    var(--rh-color-border-strong-on-light, #151515);
+  border-right: var(--rh-border-width-lg) solid var(--rh-color-border-strong-on-light);
+  border-top: var(--rh-border-width-lg) solid var(--rh-color-border-strong-on-light);
 }
 
 uxdot-sidenav-dropdown > details[open] > summary:after {
