+++ docs/assets/javascript/elements/uxdot-sidenav.css
@@ -1,0 +1,118 @@
+:host {
+  --_padding-start: var(--uxdot-sidenav-padding-start, var(--rh-space-2xl));
+  --_padding-end: var(--uxdot-sidenav-padding-end, var(--rh-space-2xl));
+  --_max-height: 100dvh;
+
+  width: 100%;
+  height: var(--_max-height);
+  top: 0;
+  z-index: var(--uxdot-sidenav-z-index, 2);
+}
+
+:host([open]) {
+  display: block !important;
+  position: fixed !important;
+}
+
+:host(:not([open])) {
+  display: none;
+}
+
+[part="close-button"] {
+  color: var(--rh-color-text-primary-on-light);
+  background-color: transparent;
+  border: none;
+  margin: 0;
+  padding: var(--rh-space-md);
+  line-height: 0 !important;
+}
+
+#container {
+  position: relative;
+  background-color: var(--rh-color-surface-lightest);
+  width: auto;
+  height: 100%;
+}
+
+#close-button-container {
+  padding-inline: var(--rh-space-md);
+  padding-block: var(--rh-space-lg);
+  max-height: var(--uxdot-masthead-max-height, 72px);
+}
+
+#close-button:focus {
+  outline: var(--rh-border-width-md) solid var(--rh-color-border-interactive-on-light);
+  border-radius: var(--rh-border-radius-default, 3px);
+}
+
+[part="overlay"] {
+  --_gray-90-rgb: var(--rh-color-gray-90-rgb);
+
+  display: none;
+  background-color: rgb(var(--_gray-90-rgb) / var(--rh-opacity-60));
+  position: fixed;
+  top: 0;
+  height: 100dvh;
+  width: 100dvw;
+  z-index: -1;
+}
+
+::slotted(ul) {
+  padding-inline: 0;
+  padding-block:
+    var(--_padding-start)
+    calc(var(--_padding-end) + (var(--rh-font-size-body-text-lg) * 1.5) + (var(--rh-space-lg) * 2));
+
+  /* allow overflow to scroll with --_padding-end + section item height */
+  list-style: none;
+  margin-block: 0 !important;
+  height: var(--_max-height);
+  overflow-y: scroll;
+  background-color: var(--rh-color-surface-lightest);
+}
+
+:host([open]) [part="overlay"] {
+  display: block;
+}
+
+@media (min-width: 320px) {
+  :host {
+    --uxdot-sidenav-width: 320px;
+
+    width: var(--uxdot-sidenav-width);
+    box-shadow: var(--rh-box-shadow-lg);
+  }
+
+  #container {
+    width: var(--uxdot-sidenav-width, 320px);
+  }
+}
+
+@media (min-width: 576px) {
+  #close-button-container {
+    padding: var(--rh-space-lg);
+  }
+}
+
+@media (min-width: 992px) {
+  :host {
+    --uxdot-siznav-z-index: 1;
+
+    position: fixed;
+    top: var(--uxdot-masthead-max-height, 72px);
+    height: calc(var(--_max-height) - var(--uxdot-masthead-max-height, 72px));
+    box-shadow: unset;
+  }
+
+  :host(:not([open])) {
+    display: block;
+  }
+
+  #close-button-container {
+    display: none;
+  }
+
+  :host([open]) [part="overlay"] {
+    display: none;
+  }
+}
